Branch HEAD b=19312
Emergency landing to fix build error on i686.

+++ lustre/tests/mpi/parallel_grouplock.c
@@ -107,13 +107,13 @@ read_buf(int fd)
                 pos = lseek(fd, 0, SEEK_CUR);
                 sprintf(errmsg, "read of file %s at pos %d for %lu bytes "
                         "returned %d: (%d) %s.\n",
-                        filename, pos, sizeof(buf), rc, errno, strerror(errno));
+                        filename, pos, (unsigned long)sizeof(buf), rc, errno, strerror(errno));
                 FAIL(errmsg);
         } else if (rc != sizeof(buf)) {
                 pos = lseek(fd, 0, SEEK_CUR);
                 sprintf(errmsg, "read of file %s at pos %d for %lu bytes "
                         "returned %d.\n",
-                        filename, pos, sizeof(buf), rc);
+                        filename, pos, (unsigned long) sizeof(buf), rc);
                 FAIL(errmsg);
         }
 }
@@ -130,12 +130,12 @@ write_buf(int fd, int index)
         if (rc == -1) {
                 sprintf(errmsg, "write of file %s at pos %d for %lu bytes "
                         "returned %d: (%d) %s.\n",
-                        filename, pos, sizeof(buf), rc, errno, strerror(errno));
+                        filename, pos, (unsigned long)sizeof(buf), rc, errno, strerror(errno));
                 FAIL(errmsg);
         } else if (rc != sizeof(buf)) {
                 sprintf(errmsg, "write of file %s at pos %d for %lu bytes "
                         "returned %d.\n",
-                        filename, pos, sizeof(buf), rc);
+                        filename, pos, (unsigned long)sizeof(buf), rc);
                 FAIL(errmsg);
         }
 }
